+++ bundle/src/main/groovy/be/orbinson/aem/groovy/console/components/Body.groovy
@@ -4,6 +4,7 @@ package be.orbinson.aem.groovy.console.components
 import be.orbinson.aem.groovy.console.GroovyConsoleService
 import be.orbinson.aem.groovy.console.audit.AuditRecord
 import be.orbinson.aem.groovy.console.audit.AuditService
+import be.orbinson.aem.groovy.console.audit.impl.AuditAccessControl
 import be.orbinson.aem.groovy.console.configuration.ConfigurationService
 import groovy.json.JsonBuilder
 import org.apache.sling.api.SlingHttpServletRequest
@@ -38,7 +39,7 @@ class Body {
         def userId = request.getParameter(USER_ID)
         def script = request.getParameter(SCRIPT)
 
-        if (script) {
+        if (script && AuditAccessControl.canAccessAuditRecord(request, userId, configurationService)) {
             auditRecord = auditService.getAuditRecord(userId, script)
         }
     }